ledgerx-web/pyproject.toml
[project]
name = "ledgerx-web"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-auth>=2.41.1",
    "google-auth-oauthlib>=1.2.2",
    "pydantic-settings>=2.11.0",
    "streamlit>=1.50.0",
]
